document2text/Parser/requirements.txt
paddleocr==2.7.0
numpy==1.26.4
pillow==10.4.0
PyMuPDF==1.20.2
opencv-python==4.6.0.66
python-docx==1.1.0
tqdm==4.66.4
ultralytics==8.2.50